+++ src/service/walletProvider/wallet.provider.ts
@@ -1,5 +1,5 @@
 import { Container, Service } from 'typedi'
-import { ITatumSdkContainer, TatumSdkWalletProviders } from "../extensions";
+import { ITatumSdkContainer, TatumSdkWalletProvider } from "../extensions";
 
 @Service({
   factory: (data: { id: string }) => {
@@ -11,7 +11,7 @@ export class WalletProvider {
   constructor(private readonly id: string) {
   }
 
-  use<T extends TatumSdkWalletProviders>(type: new (tatumSdkContainer: ITatumSdkContainer, ...args: unknown[]) => T): T {
+  use<T extends TatumSdkWalletProvider>(type: new (tatumSdkContainer: ITatumSdkContainer, ...args: unknown[]) => T): T {
     return Container.of(this.id).get(type);
   }
 }